change: wip

+++ packages/admin/pages/admin/comments.vue
@@ -1,11 +1,16 @@
 <script setup lang="ts">
 import type { Comment } from '@awesome-comment/core/types';
 import { CommentStatus } from '@awesome-comment/core/data';
+import { useAuth0 } from '@auth0/auth0-vue';
 
 type RowItem = Comment & {
-  reviewing: boolean;
+  isReviewing: boolean;
+  isReplying: boolean;
   from: string;
 }
+
+const auth0 = process.client ? useAuth0() : undefined;
+
 const start = ref<number>(0);
 const hasMore = ref<boolean>(false);
 const loadingMore = ref<boolean>(false);
@@ -21,10 +26,16 @@ definePageMeta({
 const { data, pending } = await useAsyncData(
   'comments',
   async function () {
+    if (!auth0) return;
+
+    const token = await auth0.getAccessTokenSilently();
     const { data } = await $fetch('/api/admin/comments', {
       query: {
         status: filterStatus.value === 'all' ? ['0' ,'1'] : filterStatus.value,
         start: start.value,
+      },
+      headers: {
+        Authorization: `Bearer ${token}`,
       },
     });
     hasMore.value = data?.length === 20;
@@ -32,7 +43,7 @@ const { data, pending } = await useAsyncData(
       c.user = JSON.parse((c.user || '{}') as string);
       c.status = Number(c.status);
       c.id = Number(c.id);
-      c.reviewing = false;
+      c.isReviewing = false;
       c.from = c.user_id.split('|')[ 0 ];
       return c;
     });
@@ -44,34 +55,45 @@ const { data, pending } = await useAsyncData(
   }
 );
 
-function loadMore() {
-  start.value += 20;
-}
-
 async function doReview(comment: RowItem, status: CommentStatus) {
-  comment.reviewing = true;
+  if (!auth0) return;
+
+  comment.isReviewing = true;
+  const token = await auth0.getAccessTokenSilently();
   await $fetch('/api/admin/comment/' + comment.id, {
     method: 'PATCH',
     body: {
       postId: comment.postId,
       status,
-    }
+    },
+    headers: {
+      Authorization: `Bearer ${token}`,
+    },
   });
   comment.status = status;
-  comment.reviewing = false;
+  comment.isReviewing = false;
 }
 async function doDelete(comment: RowItem, index: number): Promise<void> {
+  if (!auth0) return;
   if (!confirm('Are you sure to delete this comment?')) { return }
-  comment.reviewing = true;
+
+  comment.isReviewing = true;
+  const token = await auth0.getAccessTokenSilently();
   await $fetch('/api/admin/comment/' + comment.id, {
     method: 'DELETE',
     body: {
       postId: comment.postId,
       statue: comment.status,
     },
+    headers: {
+      Authorization: `Bearer ${token}`,
+    },
   });
   comments.value.splice(index, 1);
-  comment.reviewing = false;
+  comment.isReviewing = false;
+}
+function doLoadMore() {
+  start.value += 20;
 }
 function onFilterChange(): void {
   comments.value.length = 0;
@@ -120,40 +142,43 @@ header.flex.flex-col.mb-4.gap-4(class="sm:flex-row sm:items-center")
               v-if="comment.status === CommentStatus.Pending || comment.status === CommentStatus.Rejected"
               type="button"
               class="sm:btn-xs"
-              :disabled="comment.reviewing || loadingMore",
+              :disabled="comment.isReviewing || loadingMore",
               @click="doReview(comment, CommentStatus.Approved)"
             )
-              span.loading.loading-xs.loading-spinner(v-if="comment.reviewing")
+              span.loading.loading-xs.loading-spinner(v-if="comment.isReviewing")
               | Approve
             button.btn.btn-outline.btn-warning.btn-sm(
               v-if="comment.status === CommentStatus.Pending || comment.status === CommentStatus.Approved"
               type="button",
               class="sm:btn-xs"
-              :disabled="comment.reviewing || loadingMore",
+              :disabled="comment.isReviewing || loadingMore",
               @click="doReview(comment, CommentStatus.Rejected)"
             )
-              span.loading.loading-xs.loading-spinner(v-if="comment.reviewing")
+              span.loading.loading-xs.loading-spinner(v-if="comment.isReviewing")
               | Reject
             button.btn.btn-outline.btn-error.btn-sm(
               type="button",
               class="sm:btn-xs"
-              :disabled="comment.reviewing || loadingMore",
+              :disabled="comment.isReviewing || loadingMore",
               @click="doDelete(comment, index)"
             )
-              span.loading.loading-xs.loading-spinner(v-if="comment.reviewing")
+              span.loading.loading-xs.loading-spinner(v-if="comment.isReviewing")
               | Delete
 
+            reply-comment(
+              :comment="comment"
+            )
   button.mt-2.btn.btn-neutral.btn-sm.btn-block(
     v-if="hasMore",
     type="button",
     :disabled="loadingMore",
-    @click="loadMore",
+    @click="doLoadMore",
   )
     span.loading.loading-xs.loading-spinner(v-if="loadingMore")
     | Load More
 
   .w-full.h-32.flex.items-center.justify-center(v-if="pending")
-    .loading.loading-ring.loading-lg
+    span.loading.loading-spinner
   .w-full.h-32.flex.items-center.justify-center(v-else-if="comments.length === 0")
     .text-lg.text-center.text-neutral-content No Data to display
 </template>
